+++ src/const/auth.ts
@@ -1,6 +1,7 @@
 export const LOBE_CHAT_AUTH_HEADER = 'X-lobe-chat-auth';
 
 export const JWT_SECRET_KEY = 'LobeHub · LobeChat';
+export const NON_HTTP_PREFIX = 'http_nosafe';
 
 /* eslint-disable typescript-sort-keys/interface */
 export interface JWTPayload {
