+++ src/services/_auth.ts
@@ -1,34 +1,8 @@
-import { SignJWT, importJWK } from 'jose';
-
-import { JWTPayload, JWT_SECRET_KEY, LOBE_CHAT_AUTH_HEADER } from '@/const/auth';
+import { JWTPayload, LOBE_CHAT_AUTH_HEADER } from '@/const/auth';
 import { ModelProvider } from '@/libs/agent-runtime';
 import { useGlobalStore } from '@/store/global';
 import { modelProviderSelectors, settingsSelectors } from '@/store/global/selectors';
-
-const createJWT = async (payload: JWTPayload) => {
-  // 将AccessCode转换成适合作为密钥的格式，例如使用SHA-256进行哈希
-  const encoder = new TextEncoder();
-  const secretKey = await crypto.subtle.digest('SHA-256', encoder.encode(JWT_SECRET_KEY));
-
-  // 将密钥导入为JWK格式
-  const jwkSecretKey = await importJWK(
-    {
-      k: Buffer.from(secretKey).toString('base64'),
-      kty: 'oct',
-    },
-    'HS256',
-  );
-
-  // 获取当前时间戳
-  const now = Math.floor(Date.now() / 1000);
-
-  // 创建JWT
-  return new SignJWT(payload as Record<string, any>)
-    .setProtectedHeader({ alg: 'HS256' })
-    .setIssuedAt(now) // 设置JWT的iat（签发时间）声明
-    .setExpirationTime(now + 100) // 设置 JWT 的 exp（过期时间）为 100 s
-    .sign(jwkSecretKey);
-};
+import { createJWT } from '@/utils/jwt';
 
 const getProviderAuthPayload = (provider: string) => {
   switch (provider) {
@@ -81,7 +55,7 @@ const getProviderAuthPayload = (provider: string) => {
 const createAuthTokenWithPayload = async (payload = {}) => {
   const accessCode = settingsSelectors.password(useGlobalStore.getState());
 
-  return await createJWT({ accessCode, ...payload });
+  return await createJWT<JWTPayload>({ accessCode, ...payload });
 };
 
 interface AuthParams {
